Re-ensure that hosts are ready before rs.initiate
We've noticed an intermittent issue where sometimes the rs.initiate
doesn't work successfully. This seems to be because not all of the hosts
discovered just beforehand in the call to build_mongo_config, are ready.

This *should* be handled by the waiting that occurred in
initiate_replica.sh, but it seems that sometimes not all of the hosts
are up at that point (not all IPs are discoverable). The later
call (endpoints function) that happens in build_mongo_config then
sometimes finds more, but they're not checked to be ready.

This change performs the wait in two places: the original place (in case
there was some reason that it needed to happen before something else
there), and right before the call to rs.initiate. This should ensure
that any new hosts discovered in the build_mongo_config are ready for
replication.