Optimize Dockerfile for cache usage
Saves lots of developer time running `make` / `docker build`, by
reordering instructions to better use the cache when we make changes to
scripts, and also reduce the number of layers leading to a smaller time
to build.

For instance, a developer might run the command below multiple times in
a row while making changes to the scripts that are copied into the
image:

    make VERSION=3.2 SKIP_SQUASH=1

We've moving informative instructions close together at the top, while
moving `ADD root /` down as much as possible, so that changes to the
scripts invalidate less cached layers.